+++ packages/scaffold/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @web3modal/scaffold
 
+## 4.1.10
+
+### Patch Changes
+
+- Fix activity list styling issue
+
+- Updated dependencies []:
+  - @web3modal/common@4.1.10
+  - @web3modal/core@4.1.10
+  - @web3modal/siwe@4.1.10
+  - @web3modal/ui@4.1.10
+  - @web3modal/wallet@4.1.10
+
 ## 4.1.9
 
 ### Patch Changes